Time: 110 ms (99.26%) | Memory: 48.3 MB (98.92%) - LeetSync

+++ 2711-minimum-time-to-visit-a-cell-in-a-grid/minimum-time-to-visit-a-cell-in-a-grid.cpp
@@ -10,7 +10,7 @@ public:
         priority_queue<pair<int, int>, vector<pair<int, int>>, greater<>> pq;
         pq.push({0, 0});
         // Directions: {right, down, left, up}
-        short d[5] = {0, 1, 0, -1, 0};
+        int d[5] = {0, 1, 0, -1, 0};
         int waitTime,i;
         while (!pq.empty()) {
             auto [currentDist, cell] = pq.top();
@@ -23,7 +23,7 @@ public:
             if (dist[cell]<currentDist) continue;
             i = 0;
             while(i < 4) {
-                short nx = x + d[i], ny = y + d[++i];
+                int nx = x + d[i], ny = y + d[++i];
                 if (nx < 0 || ny < 0 || nx >= m || ny >= n)
                     continue;
                 int c_pos = nx + ny * m;
